+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
-## <a id="v1.10.6"></a>[v1.10.6] - 2024-03-11
+## <a id="v1.10.7"></a>[v1.10.7] - 2024-03-12
 
 ### Added
 
@@ -19,6 +19,8 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - engine leak from analysis worker
+
+## <a id="v1.10.6"></a>[v1.10.6] - 2024-03-11 [YANKED]
 
 ## <a id="v1.10.5"></a>[v1.10.5] - 2024-02-22